backend/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
black==21.6b0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.1
colorama==0.4.4
DateTime==5.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==3.14.0
django-recaptcha==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==10.0.0
platformdirs==3.1.0
PyJWT==2.7.0
pytz==2023.3
regex==2021.4.4
requests==2.31.0
sqlparse==0.4.4
toml==0.10.2
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.20.0
zope.interface==6.0
